this "when to use section" is misleading
I think a when to use section would be nice to have, but the current one only makes sense as a set of restrictions applied to fold, not reducers in generally. 

If you look at http://clojure.com/blog/2012/05/08/reducers-a-library-and-model-for-collection-processing.html under the heading "More Opportunity (i.e. Work)" rhickey explicitly called out reducers as a mechanism for solving "the lazily dangling open resource problem"