Replace "+" with "addition"
It looks like asciidoc treats a plus sign at the end of a line specially (http://asciidoctor.org/docs/asciidoc-writers-guide/#wrapped-text-and-hard-line-breaks), so the plus isn't showing up on the web page (https://clojure.org/guides/learn/syntax). Regardless, it seems less confusing to use the full word "addition", rather than using "+" to refer to both the var and the concept of addition in the same sentence.