Explicit example of 1-arity return type hint
To an experienced Clojure user it is clear that `(defn hinted-single ^String [])` is equivalent to `(defn hinted-single (^String []))` but it isn't so obvious to somebody less experienced, especially one used to the fact that function and type declaration is "magic" and not simply a reuse of mechanisms used in user code (namely metadata and metadata reader macro). So providing an explicit example will help them a lot to relate it to their code (that rarely uses multiple arities).