Remove spaces in companies CSV file
The companies CSV file has a space after the comma separator, making it
incompatible with org.clojure/data.csv. data.csv will read the URL cell as if it
has a leading space character. On top of that it will fail to detect a quoted
cell in the URL column, because the first character after the separator is a
space, rather than a quote.

This is a step in preparation of merging in data from a crowdsourced spreadsheet
of Clojure companies.