Link Leiningen and Boot in the Windows section
Whilst they are both linked in the Build Tools section at the bottom of the page, Windows users might not read that far. Linking them where they are mentioned in the Windows section will make it easier for Windows beginners to navigate to the appropriate tools, until the PowerShell version of the CLI tools is available.