Update URL scheme on Functions doc
It turns out that when using http://www.w3.org/ the connection will
timeout, but when using https://www.w3.org/ results are as expected. I
can reproduce this using cURL, so it's not a Clojure thing, but still,
it could be frustrating to newbies to get an error while trying to
work on that exercise.