Add "Get Programming with Clojure" to the books list
The book is in early preview. 5 out 14 Units are already published. 
The 5 units cover the basics of the language, form S-expression syntax to function definition and instructions to launch a Clojure program from the command line. 
I believe the early preview version of the book is valuable for Clojure beginners.