Update cheatsheet for Clojure 1.10
Here is what I have already added that is new in Clojure 1.10:

Added to Namespace / From symbol section:
    requiring-resolve
Added to Java Interoperation / Exceptions section:
    ex-cause ex-message clojure.main/ex-str

And here are the new things in Clojure 1.10 that have not been added.
If someone would like them added to the cheatsheet, please recommend a
category (or multiple categories) to add them to, whether existing or
new.

    PrintWriter-on
    read+string
    add-tap
    remove-tap
    tap>
    clojure.core.protocols/Datafiable
    clojure.core.protocols/Navigable
    clojure.core.protocols/datafy
    clojure.core.protocols/nav
    clojure.core.server/io-prepl
    clojure.core.server/prepl
    clojure.core.server/remote-prepl
    clojure.core.specs.alpha/even-number-of-forms?
    clojure.spec.gen.alpha/shuffle
    clojure.datafy/datafy
    clojure.datafy/nav
    clojure.main/renumbering-read