Update recommendations for VS Code and Atom
I've updated the recommended plugins/extensions for Atom, to Chlorine, and for VS Code, to Calva. In the latter case I am quite biased, but as [this question](https://github.com/avli/clojureVSCode/issues/113) suggests, `clojureVSCode` does not fully qualify the expectations for an editor integration as they are expressed on this page. Calva does, however.