Update Clojurians Slack link
Right now, the link is to clojurians.net which used to be the "inviter app" but was changed to just redirect to a join-Slack invitation link (because Slack changed how invitations worked). Unfortunately, that link has expired and the Clojurians Admin team do not have access to clojurians.net to change the link to a new one.

So I'm proposing changing the link on clojure.org to directly be a new invitation link -- it should not expire until 2,000 people have used it (in theory) -- and once we get control of clojurians.net again, I'll submit a PR to change this page back to clojurians.net. I'll also keep an eye on signups and if it looks like this new link will "run out", I'll get a new one created and submit another PR.

Sorry for the churn but we're out of options for the time being since Slack has no self-signup option.