Remove broken clojure-doc.org link
Since clojure-doc.org was resurrected as GitHub pages, the URLs changed and this link was broken. Since the Emacs page on clojure-doc.org is going away and I already submitted a PR for the editors page on clojure.org, I think it makes sense to just remove this old link instead of trying to replace it with something else.