Weird characters: add names
The "weird characters"/"Reading Clojure Characters" page's goal is to improve
findability. Adding English names for these special characters makes it easier
to search within a page as well as to use search engines (which can make it
difficult to search with special characters). Referring to these characters by
name is also helpful for Clojure users for whom English is not their native
language.